# backend/app/api/v1/public.py
from logging import getLogger

from flask import Blueprint, request

from app.api.utils.validation import parse, parse_query
from app.core.extensions import auth
from app.db.context import get_core_db, get_response_db
from app.openapi import openapi_route
from app.schema.api.requests.public_links import ResolveTokenRequest
from app.schema.api.requests.submissions.create import (
    LinkSubmissionRequest,
    SlugSubmissionRequest,
)
from app.schema.api.requests.surveys import ListPublicSurveysRequest
from app.schema.api.responses.public_links import PublicLinkResponses, ResolveLinkResponses
from app.schema.api.responses.submissions import AnswerResponses, CoreSubmissionResponses, LinkedSubmissionResponses
from app.schema.api.responses.surveys import (
    PaginatedPublicSurveysResponses,
    PublicSurveyResponses,
    SurveyResponses,
    SurveyVersionResponses,
)
from app.services.public_links import SurveyLinkService
from app.services.public_surveys import PublicSurveyService
from app.services.submissions import SubmissionIntakeService
from app.services.users import UserService

logger = getLogger(__name__)

public_bp = Blueprint("public_v1", __name__)

users_service = UserService()
survey_link_service = SurveyLinkService()
submission_intake_service = SubmissionIntakeService()
public_survey_service = PublicSurveyService()


@openapi_route(
    summary="List public surveys",
    query_model=ListPublicSurveysRequest,
    response_model=PaginatedPublicSurveysResponses,
    tags=["Public"],
    auth_required=False,
)
@public_bp.route("/surveys", methods=["GET"])
def list_public_surveys():
    params = parse_query(ListPublicSurveysRequest, request)
    core_db = get_core_db()

    result = public_survey_service.list_public_surveys(core_db, page=params.page, page_size=params.page_size)
    response = PaginatedPublicSurveysResponses(
        items=[SurveyResponses.model_validate(s) for s in result.surveys],
        total=result.total,
        page=result.page,
        page_size=result.page_size,
    )
    return response.model_dump(mode="json"), 200


@openapi_route(
    summary="Get public survey",
    response_model=PublicSurveyResponses,
    tags=["Public"],
    auth_required=False,
)
@public_bp.route("/surveys/<public_slug:public_slug>", methods=["GET"])
def get_public_survey(public_slug: str):
    core_db = get_core_db()

    result = public_survey_service.get_public_survey(core_db, public_slug=public_slug)
    response = PublicSurveyResponses(
        survey=SurveyResponses.model_validate(result.survey),
        published_version=(
            SurveyVersionResponses.model_validate(result.published_version) if result.published_version else None
        ),
    )
    return response.model_dump(mode="json"), 200


@openapi_route(
    summary="Resolve survey link",
    query_model=ResolveTokenRequest,
    response_model=ResolveLinkResponses,
    tags=["Public Links"],
    auth="optional",
    description=(
        "Anonymous access is allowed for links that do not require authentication. "
        "If the resolved link requires authentication, a bearer token must be supplied."
    ),
)
@public_bp.route("/links/resolve", methods=["GET"])
@auth.optional_auth()
def resolve_link():
    payload = parse_query(ResolveTokenRequest, request)

    core_db = get_core_db()

    current_sub = auth.get_optional_current_user_sub()
    user = users_service.get_user_by_sub(db=core_db, auth0_user_id=current_sub) if current_sub is not None else None
    result = survey_link_service.resolve_link(core_db, payload=payload, actor=user)
    response = ResolveLinkResponses(
        link=PublicLinkResponses.model_validate(result.link),
        survey=SurveyResponses.model_validate(result.survey),
        published_version=SurveyVersionResponses.model_validate(result.published_version),
    )

    return response.model_dump(mode="json"), 200


@openapi_route(
    summary="Create slug submission",
    request_model=SlugSubmissionRequest,
    response_model=LinkedSubmissionResponses,
    status_code=201,
    tags=["Public Submissions"],
    auth="optional",
)
@public_bp.route("/submissions/slug", methods=["POST"])
@auth.optional_auth()
def create_slug_submission():
    payload = parse(SlugSubmissionRequest, request)
    core_db = get_core_db()
    response_db = get_response_db()
    submitted_by_user_id = None
    current_sub = auth.get_optional_current_user_sub()
    if current_sub is not None:
        submitted_by_user_id = users_service.get_user_by_sub(db=core_db, auth0_user_id=current_sub).id

    linked = submission_intake_service.create_slug_submission(
        core_db,
        response_db,
        payload=payload,
        submitted_by_user_id=submitted_by_user_id,
    )
    response = LinkedSubmissionResponses(
        core=CoreSubmissionResponses.model_validate(linked.core_submission),
        answers=[AnswerResponses.model_validate(a) for a in linked.answers],
    )
    return response.model_dump(mode="json"), 201


@openapi_route(
    summary="Create link submission",
    request_model=LinkSubmissionRequest,
    response_model=LinkedSubmissionResponses,
    status_code=201,
    tags=["Public Submissions"],
    auth="optional",
)
@public_bp.route("/submissions/link", methods=["POST"])
@auth.optional_auth()
def create_link_submission():
    payload = parse(LinkSubmissionRequest, request)
    core_db = get_core_db()
    response_db = get_response_db()
    current_sub = auth.get_optional_current_user_sub()
    user = users_service.get_user_by_sub(db=core_db, auth0_user_id=current_sub) if current_sub is not None else None
    linked = submission_intake_service.create_link_submission(
        core_db,
        response_db,
        payload=payload,
        actor=user,
    )
    response = LinkedSubmissionResponses(
        core=CoreSubmissionResponses.model_validate(linked.core_submission),
        answers=[AnswerResponses.model_validate(a) for a in linked.answers],
    )
    return response.model_dump(mode="json"), 201