# backend/openapi.yaml
openapi: 3.1.1
info:
  description: REST API for the FlowForm survey platform.
  contact:
    name: FlowForm Support
    email: support@flow-form.com.au
  title: FlowForm API
  version: 0.1.0
servers:
- url: http://localhost:5000
  description: Local development
security:
- BearerAuth: []
...
paths:
...
  /api/v1/public/surveys:
    get:
      operationId: listPublicSurveys
      summary: List public surveys
      description: List public surveys
      tags:
      - Public
      security: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - name: page_size
        in: query
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Page Size
          type: integer
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPublicSurveysResponses'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/public/surveys/{public_slug}:
    get:
      operationId: getPublicSurvey
      summary: Get public survey
      description: Get public survey
      tags:
      - Public
      security: []
      parameters:
      - name: public_slug
        in: path
        required: true
        schema:
          type: string
          maxLength: 80
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSurveyResponses'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/public/links/resolve:
    get:
      operationId: resolveLink
      summary: Resolve survey link
      description: Anonymous access is allowed for links that do not require authentication. If the resolved link requires
        authentication, a bearer token must be supplied.
      tags:
      - Public Links
      security:
      - {}
      - BearerAuth: []
      parameters:
      - name: token
        in: query
        required: true
        schema:
          maxLength: 256
          minLength: 1
          title: Token
          type: string
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolveLinkResponses'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/public/submissions/slug:
    post:
      operationId: createSlugSubmission
      summary: Create slug submission
      description: Create slug submission
      tags:
      - Public Submissions
      security:
      - {}
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlugSubmissionRequest'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkedSubmissionResponses'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/public/submissions/link:
    post:
      operationId: createLinkSubmission
      summary: Create link submission
      description: Create link submission
      tags:
      - Public Submissions
      security:
      - {}
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkSubmissionRequest'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkedSubmissionResponses'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    PaginatedPublicSurveysResponses:
      description: API response shape for a paginated list of public surveys.
      properties:
        items:
          items:
            $ref: '#/components/schemas/SurveyResponses'
          maxItems: 100
          title: Items
          type: array
        total:
          title: Total
          type: integer
        page:
          title: Page
          type: integer
        page_size:
          title: Page Size
          type: integer
      required:
      - items
      - total
      - page
      - page_size
      title: PaginatedPublicSurveysResponses
      type: object
    PublicSurveyResponses:
      description: API response shape for a publicly accessible survey with its published version.
      properties:
        survey:
          $ref: '#/components/schemas/SurveyResponses'
        published_version:
          anyOf:
          - $ref: '#/components/schemas/SurveyVersionResponses'
          - type: 'null'
          default: null
      required:
      - survey
      title: PublicSurveyResponses
      type: object
    ResolveLinkResponses:
      description: API response shape for resolving a public link token.
      properties:
        link:
          $ref: '#/components/schemas/PublicLinkResponses'
        survey:
          anyOf:
          - $ref: '#/components/schemas/SurveyResponses'
          - type: 'null'
          default: null
        published_version:
          anyOf:
          - $ref: '#/components/schemas/SurveyVersionResponses'
          - type: 'null'
          default: null
      required:
      - link
      title: ResolveLinkResponses
      type: object
...
  responses:
    BadRequestError:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_email_claim:
              summary: MissingEmailClaimError
              value:
                code: MISSING_EMAIL_CLAIM
                message: ID token did not contain an email claim.
            survey_no_response_store:
              summary: SurveyNoResponseStoreError
              value:
                code: INVALID_REQUEST
                message: <example>
            submission_invalid:
              summary: SubmissionInvalidError
              value:
                code: INVALID_SUBMISSION
                message: Submission is invalid.
            submission_invalid_timestamps:
              summary: SubmissionInvalidTimestampsError
              value:
                code: INVALID_SUBMISSION_TIMESTAMPS
                message: Submission timestamps are invalid.
            submission_answers_required:
              summary: SubmissionAnswersRequiredError
              value:
                code: SUBMISSION_ANSWERS_REQUIRED
                message: At least one answer is required.
    UnauthorizedError:
      description: Authentication required or token invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_id_token_subject:
              summary: InvalidIdTokenSubjectError
              value:
                code: INVALID_ID_TOKEN
                message: ID token did not contain a valid subject.
            token_subject_mismatch:
              summary: TokenSubjectMismatchError
              value:
                code: TOKEN_SUBJECT_MISMATCH
                message: ID token subject did not match the access token subject.
            link_auth_required:
              summary: LinkAuthRequiredError
              value:
                code: LINK_AUTH_REQUIRED
                message: This link requires authentication.
    ForbiddenError:
      description: Authenticated but not authorized for this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              summary: ForbiddenError
              value:
                code: FORBIDDEN
                message: You do not have permission to access this resource.
            link_inactive:
              summary: LinkInactiveError
              value:
                code: LINK_INACTIVE
                message: This link is inactive
            link_expired:
              summary: LinkExpiredError
              value:
                code: LINK_EXPIRED
                message: This link has expired
            link_no_response:
              summary: LinkNoResponseError
              value:
                code: LINK_NO_RESPONSE
                message: This link does not allow submissions
            link_assignment_mismatch:
              summary: LinkAssignmentMismatchError
              value:
                code: LINK_ASSIGNED_TO_ANOTHER_USER
                message: This link is assigned to a different user.
            link_already_used:
              summary: LinkAlreadyUsedError
              value:
                code: LINK_ALREADY_USED
                message: This link has already been used.
            survey_not_accessible:
              summary: SurveyNotAccessibleError
              value:
                code: SURVEY_NOT_ACCESSIBLE
                message: This survey is not accessible in this context.
            member_self_action:
              summary: MemberSelfActionError
              value:
                code: MEMBER_SELF_ACTION
                message: You cannot modify your own membership.
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            link_not_found:
              summary: LinkNotFoundError
              value:
                code: NOT_FOUND
                message: Invalid or unknown token
            public_link_not_found:
              summary: PublicLinkNotFoundError
              value:
                code: NOT_FOUND
                message: Survey link not found
            survey_not_found:
              summary: SurveyNotFoundError
              value:
                code: SURVEY_NOT_FOUND
                message: Survey 0 was not found in project 0.
            survey_not_published:
              summary: SurveyNotPublishedError
              value:
                code: SURVEY_NOT_PUBLISHED
                message: Survey 0 in project 0 is not published.
            survey_not_found_by_slug:
              summary: SurveyNotFoundBySlugError
              value:
                code: NOT_FOUND
                message: Survey not found
            project_not_found:
              summary: ProjectNotFoundError
              value:
                code: NOT_FOUND
                message: Project id=None not found.
            version_not_found:
              summary: VersionNotFoundError
              value:
                code: NOT_FOUND
                message: Version 0 was not found in survey 0.
            node_not_found:
              summary: NodeNotFoundError
              value:
                code: NOT_FOUND
                message: Node not found
            scoring_rule_not_found:
              summary: ScoringRuleNotFoundError
              value:
                code: NOT_FOUND
                message: Scoring rule not found
            submission_not_found:
              summary: SubmissionNotFoundError
              value:
                code: SUBMISSION_NOT_FOUND
                message: Submission 0 not found.
            user_not_found:
              summary: UserNotFoundError
              value:
                code: USER_NOT_FOUND
                message: User <example> not found.
            invitation_not_found:
              summary: InvitationNotFoundError
              value:
                code: NOT_FOUND
                message: Invitation not found.
            project_role_not_found:
              summary: ProjectRoleNotFoundError
              value:
                code: NOT_FOUND
                message: Project role not found.
            member_not_found:
              summary: MemberNotFoundError
              value:
                code: NOT_FOUND
                message: Member not found.
            survey_role_not_found:
              summary: SurveyRoleNotFoundError
              value:
                code: NOT_FOUND
                message: Survey role not found.
            survey_member_role_not_found:
              summary: SurveyMemberRoleNotFoundError
              value:
                code: NOT_FOUND
                message: Survey member role assignment not found.
    ConflictError:
      description: Conflict with the current resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            survey_delete_published:
              summary: SurveyDeletePublishedError
              value:
                code: SURVEY_HAS_PUBLISHED_VERSION
                message: Cannot delete survey 0 because it has a published version.
            survey_publish:
              summary: SurveyPublishError
              value:
                code: SURVEY_PUBLISH_ERROR
                message: <example>
                details:
                  validation_errors: <example>
            project_slug_conflict:
              summary: ProjectSlugConflictError
              value:
                code: CONFLICT
                message: Conflict — a project with that slug already exists
            user_bootstrap_conflict:
              summary: UserBootstrapConflictError
              value:
                code: CONFLICT
                message: Conflict — a user with email '<example>' already exists
            survey_slug_conflict:
              summary: SurveySlugConflictError
              value:
                code: CONFLICT
                message: Conflict — check slug uniqueness
            version_already_archived:
              summary: VersionAlreadyArchivedError
              value:
                code: VERSION_ALREADY_ARCHIVED
                message: Version is already archived
            version_not_published:
              summary: VersionNotPublishedError
              value:
                code: VERSION_NOT_PUBLISHED
                message: Version is not published
            version_is_active_published:
              summary: VersionIsActivePublishedError
              value:
                code: VERSION_IS_ACTIVE_PUBLISHED
                message: Cannot archive the active published version directly; unpublish it first
            version_not_editable:
              summary: VersionNotEditableError
              value:
                code: VERSION_NOT_EDITABLE
                message: Version is '<example>' — content can only be edited on draft versions
            content_key_conflict:
              summary: ContentKeyConflictError
              value:
                code: CONFLICT
                message: <example>
            invitation_already_exists:
              summary: InvitationAlreadyExistsError
              value:
                code: INVITATION_EXISTS
                message: A pending invitation already exists for <example>.
            invitation_not_pending:
              summary: InvitationNotPendingError
              value:
                code: INVITATION_NOT_PENDING
                message: This invitation is no longer pending.
            already_amember:
              summary: AlreadyAMemberError
              value:
                code: ALREADY_A_MEMBER
                message: This user is already a member of the project.
            project_role_system_protected:
              summary: ProjectRoleSystemProtectedError
              value:
                code: ROLE_SYSTEM_PROTECTED
                message: System roles cannot be modified or deleted.
            member_owner_protected:
              summary: MemberOwnerProtectedError
              value:
                code: MEMBER_OWNER_PROTECTED
                message: Owner roles cannot be assigned, reassigned, removed, or suspended.
            survey_role_name_conflict:
              summary: SurveyRoleNameConflictError
              value:
                code: CONFLICT
                message: A survey role with that name already exists.
            survey_member_role_already_assigned:
              summary: SurveyMemberRoleAlreadyAssignedError
              value:
                code: SURVEY_MEMBER_ROLE_CONFLICT
                message: This member already has a survey role assigned for this survey.
            users_email_key:
              summary: 'DB rule: EMAIL_CONFLICT'
              value:
                code: EMAIL_CONFLICT
                message: An account with email '<email>' already exists.
            users_auth0_user_id_key:
              summary: 'DB rule: AUTH0_ID_CONFLICT'
              value:
                code: AUTH0_ID_CONFLICT
                message: Auth0 user id '<auth0_user_id>' is already associated with another account.
            projects_slug_key:
              summary: 'DB rule: PROJECT_SLUG_CONFLICT'
              value:
                code: PROJECT_SLUG_CONFLICT
                message: Project slug '<slug>' is already in use.
            uq_project_invitations_pending_project_email:
              summary: 'DB rule: INVITATION_EXISTS'
              value:
                code: INVITATION_EXISTS
                message: A pending invitation already exists for '<invited_email>' in project id=<project_id>.
            fk_project_invitations_role_same_project:
              summary: 'DB rule: INVITATION_ROLE_MISMATCH'
              value:
                code: INVITATION_ROLE_MISMATCH
                message: Role id=<role_id> does not belong to project id=<project_id>.
            surveys_public_slug_key:
              summary: 'DB rule: SURVEY_SLUG_CONFLICT'
              value:
                code: SURVEY_SLUG_CONFLICT
                message: Survey slug '<public_slug>' is already in use.
            fk_surveys_default_store:
              summary: 'DB rule: SURVEY_STORE_MISSING'
              value:
                code: SURVEY_STORE_MISSING
                message: Default response store id=<default_response_store_id> does not exist.
            fk_surveys_default_store_same_project:
              summary: 'DB rule: SURVEY_STORE_PROJECT_CONFLICT'
              value:
                code: SURVEY_STORE_PROJECT_CONFLICT
                message: Default response store id=<default_response_store_id> does not belong to project id=<project_id>.
            fk_surveys_published_version_same_survey:
              summary: 'DB rule: SURVEY_PUBLISHED_VERSION_MISMATCH'
              value:
                code: SURVEY_PUBLISHED_VERSION_MISMATCH
                message: Published version id=<published_version_id> does not belong to survey id=<survey_id>.
            published_version_id_must_reference_a_published_non_deleted_version_of_the_same_survey:
              summary: 'DB rule: SURVEY_PUBLISHED_VERSION_INVALID'
              value:
                code: SURVEY_PUBLISHED_VERSION_INVALID
                message: Published version id=<published_version_id> must be a published, non-deleted version of survey id=<survey_id>.
            uq_survey_versions_survey_id_version_number:
              summary: 'DB rule: VERSION_NUMBER_CONFLICT'
              value:
                code: VERSION_NUMBER_CONFLICT
                message: Survey already has version_number=<version_number>.
            uq_survey_versions_one_published:
              summary: 'DB rule: VERSION_PUBLISHED_CONFLICT'
              value:
                code: VERSION_PUBLISHED_CONFLICT
                message: Survey id=<survey_id> already has a published version.
            cannot_delete_the_active_published_survey_version_while_it_is_referenced_by_surveys_published_version_id:
              summary: 'DB rule: VERSION_DELETE_PROTECTED'
              value:
                code: VERSION_DELETE_PROTECTED
                message: Published version id=<survey_version_id> cannot be deleted.
            cannot_unpublish_or_soft_delete_the_active_published_survey_version_while_it_is_referenced_by_surveys_published_version_id:
              summary: 'DB rule: VERSION_STATE_PROTECTED'
              value:
                code: VERSION_STATE_PROTECTED
                message: Published version id=<survey_version_id> cannot be unpublished or deleted.
            uq_survey_links_token_hash:
              summary: 'DB rule: LINK_TOKEN_CONFLICT'
              value:
                code: LINK_TOKEN_CONFLICT
                message: This survey link token is already in use.
            uq_survey_links_survey_id_token_prefix:
              summary: 'DB rule: LINK_PREFIX_CONFLICT'
              value:
                code: LINK_PREFIX_CONFLICT
                message: Survey id=<survey_id> already has a link with this token prefix.
            uq_survey_roles_project_id_name:
              summary: 'DB rule: SURVEY_ROLE_NAME_CONFLICT'
              value:
                code: SURVEY_ROLE_NAME_CONFLICT
                message: Project already has a survey role named '<name>'.
            uq_response_subject_mappings_project_id_user_id:
              summary: 'DB rule: SUBJECT_USER_CONFLICT'
              value:
                code: SUBJECT_USER_CONFLICT
                message: Project id=<project_id> already has a subject mapping for user id=<user_id>.
            uq_response_subject_mappings_project_id_pseudonymous_subject_id:
              summary: 'DB rule: SUBJECT_ID_CONFLICT'
              value:
                code: SUBJECT_ID_CONFLICT
                message: Project id=<project_id> already uses pseudonymous subject id=<pseudonymous_subject_id>.
            fk_survey_submissions_survey_same_project:
              summary: 'DB rule: SUBMISSION_SURVEY_PROJECT_CONFLICT'
              value:
                code: SUBMISSION_SURVEY_PROJECT_CONFLICT
                message: Survey id=<survey_id> does not belong to project id=<project_id>.
            fk_survey_submissions_version_same_survey:
              summary: 'DB rule: SUBMISSION_VERSION_SURVEY_CONFLICT'
              value:
                code: SUBMISSION_VERSION_SURVEY_CONFLICT
                message: Survey version id=<survey_version_id> does not belong to survey id=<survey_id>.
            fk_survey_submissions_store_same_project:
              summary: 'DB rule: SUBMISSION_STORE_PROJECT_CONFLICT'
              value:
                code: SUBMISSION_STORE_PROJECT_CONFLICT
                message: Response store id=<response_store_id> does not belong to project id=<project_id>.
            fk_survey_submissions_survey_link_same_survey:
              summary: 'DB rule: SUBMISSION_LINK_SURVEY_CONFLICT'
              value:
                code: SUBMISSION_LINK_SURVEY_CONFLICT
                message: Survey link id=<survey_link_id> does not belong to survey id=<survey_id>.
            fk_survey_submissions_subject_same_project:
              summary: 'DB rule: SUBMISSION_SUBJECT_PROJECT_CONFLICT'
              value:
                code: SUBMISSION_SUBJECT_PROJECT_CONFLICT
                message: Pseudonymous subject id='<pseudonymous_subject_id>' does not belong to project id=<project_id>.
            uq_survey_submissions_external_submission_id:
              summary: 'DB rule: SUBMISSION_EXTERNAL_ID_CONFLICT'
              value:
                code: SUBMISSION_EXTERNAL_ID_CONFLICT
                message: External submission id='<external_submission_id>' is already in use for response_store_id=<response_store_id>.
            uq_survey_questions_survey_version_id_question_key:
              summary: 'DB rule: RULE_KEY_CONFLICT'
              value:
                code: RULE_KEY_CONFLICT
                message: Survey version id=<survey_version_id> already uses rule_key (node id=<survey_question_id>).
            uq_survey_questions_survey_version_id_sort_key:
              summary: 'DB rule: SORT_KEY_CONFLICT'
              value:
                code: SORT_KEY_CONFLICT
                message: Survey version id=<survey_version_id> already uses that sort_key.
            cannot_modify_components_of_a_published_survey_version:
              summary: 'DB rule: SCORING_VERSION_LOCKED'
              value:
                code: SCORING_VERSION_LOCKED
                message: Scoring rules cannot be changed on a published survey version.
            uq_survey_scoring_rules_survey_version_id_scoring_key:
              summary: 'DB rule: SCORING_KEY_CONFLICT'
              value:
                code: SCORING_KEY_CONFLICT
                message: Survey version id=<survey_version_id> already uses scoring_key='<scoring_key>'.
            uq_project_roles_project_name:
              summary: 'DB rule: PROJECT_ROLE_NAME_CONFLICT'
              value:
                code: PROJECT_ROLE_NAME_CONFLICT
                message: Project already has a role named '<name>'.
            uq_project_memberships_user_project:
              summary: 'DB rule: PROJECT_MEMBERSHIP_CONFLICT'
              value:
                code: PROJECT_MEMBERSHIP_CONFLICT
                message: User id=<user_id> is already a member of project id=<project_id>.
            fk_project_memberships_role_same_project:
              summary: 'DB rule: PROJECT_ROLE_MISMATCH'
              value:
                code: PROJECT_ROLE_MISMATCH
                message: Role id=<role_id> does not belong to project id=<project_id>.
            fk_survey_membership_roles_survey_same_project:
              summary: 'DB rule: SURVEY_MEMBERSHIP_SURVEY_MISMATCH'
              value:
                code: SURVEY_MEMBERSHIP_SURVEY_MISMATCH
                message: The survey does not belong to the given project.
            fk_survey_membership_roles_membership_same_project:
              summary: 'DB rule: SURVEY_MEMBERSHIP_PROJECT_MISMATCH'
              value:
                code: SURVEY_MEMBERSHIP_PROJECT_MISMATCH
                message: The membership does not belong to the given project.
            fk_survey_membership_roles_role_same_project:
              summary: 'DB rule: SURVEY_ROLE_PROJECT_MISMATCH'
              value:
                code: SURVEY_ROLE_PROJECT_MISMATCH
                message: The survey role does not belong to the given project.
...
    RateLimitError:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit_exceeded:
              summary: RateLimitExceededError
              value:
                code: RATE_LIMIT_EXCEEDED
                message: Rate limit exceeded. Please retry after some time.
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Auth0-issued JWT access token. Send as `Authorization: Bearer <token>`.'